+++ demo/start.js
@@ -51,8 +51,7 @@ const clientConfig = {
     extensions: [ '.tsx', '.ts', '.js' ],
     alias: {
       common: path.resolve('./out/common'),
-      core: path.resolve('./out/core'),
-      ui: path.resolve('./out/ui')
+      browser: path.resolve('./out/browser')
     }
   },
   output: {
