+++ src/AccessibilityManager.ts
@@ -6,10 +6,10 @@
 import * as Strings from './Strings';
 import { ITerminal, IBuffer } from './Types';
 import { isMac } from 'common/Platform';
-import { RenderDebouncer } from 'ui/RenderDebouncer';
-import { addDisposableDomListener } from 'ui/Lifecycle';
+import { RenderDebouncer } from 'browser/RenderDebouncer';
+import { addDisposableDomListener } from 'browser/Lifecycle';
 import { Disposable } from 'common/Lifecycle';
-import { ScreenDprMonitor } from 'ui/ScreenDprMonitor';
+import { ScreenDprMonitor } from 'browser/ScreenDprMonitor';
 import { IRenderDimensions } from './renderer/Types';
 
 const MAX_ROWS_TO_READ = 20;
