+++ src/MouseZoneManager.ts
@@ -5,7 +5,7 @@
 
 import { ITerminal, IMouseZoneManager, IMouseZone } from './Types';
 import { Disposable } from 'common/Lifecycle';
-import { addDisposableDomListener } from 'ui/Lifecycle';
+import { addDisposableDomListener } from 'browser/Lifecycle';
 
 const HOVER_DURATION = 500;
