+++ src/Terminal.ts
@@ -35,7 +35,7 @@ import { Renderer } from './renderer/Renderer';
 import { Linkifier } from './Linkifier';
 import { SelectionManager } from './SelectionManager';
 import * as Browser from 'common/Platform';
-import { addDisposableDomListener } from 'ui/Lifecycle';
+import { addDisposableDomListener } from 'browser/Lifecycle';
 import * as Strings from './Strings';
 import { MouseHelper } from './MouseHelper';
 import { SoundManager } from './SoundManager';
@@ -49,12 +49,12 @@ import { evaluateKeyboardEvent } from 'common/input/Keyboard';
 import { EventEmitter2, IEvent } from 'common/EventEmitter2';
 import { Attributes, DEFAULT_ATTR_DATA } from 'common/buffer/BufferLine';
 import { applyWindowsMode } from './WindowsMode';
-import { ColorManager } from 'ui/ColorManager';
+import { ColorManager } from 'browser/ColorManager';
 import { RenderCoordinator } from './renderer/RenderCoordinator';
 import { IOptionsService } from 'common/services/Services';
 import { OptionsService } from 'common/services/OptionsService';
-import { ICharSizeService } from 'ui/services/Services';
-import { CharSizeService } from 'ui/services/CharSizeService';
+import { ICharSizeService } from 'browser/services/Services';
+import { CharSizeService } from 'browser/services/CharSizeService';
 
 // Let it work inside Node.js for automated testing purposes.
 const document = (typeof window !== 'undefined') ? window.document : null;
