+++ src/browser/Lifecycle.ts
