+++ src/browser/RenderDebouncer.ts
