+++ src/browser/services/CharSizeService.ts
@@ -5,7 +5,7 @@
 
 import { IOptionsService } from 'common/services/Services';
 import { IEvent, EventEmitter2 } from 'common/EventEmitter2';
-import { ICharSizeService } from 'ui/services/Services';
+import { ICharSizeService } from 'browser/services/Services';
 
 export class CharSizeService implements ICharSizeService {
   public width: number = 0;
