+++ src/renderer/Renderer.ts
@@ -11,8 +11,8 @@ import { ITerminal, CharacterJoinerHandler } from '../Types';
 import { LinkRenderLayer } from './LinkRenderLayer';
 import { CharacterJoinerRegistry } from '../renderer/CharacterJoinerRegistry';
 import { Disposable } from 'common/Lifecycle';
-import { IColorSet } from 'ui/Types';
-import { ICharSizeService } from 'ui/services/Services';
+import { IColorSet } from 'browser/Types';
+import { ICharSizeService } from 'browser/services/Services';
 
 export class Renderer extends Disposable implements IRenderer {
   private _renderLayers: IRenderLayer[];
