+++ src/renderer/dom/DomRenderer.ts
@@ -8,8 +8,8 @@ import { ILinkifierEvent, ITerminal, CharacterJoinerHandler } from '../../Types'
 import { BOLD_CLASS, ITALIC_CLASS, CURSOR_CLASS, CURSOR_STYLE_BLOCK_CLASS, CURSOR_BLINK_CLASS, CURSOR_STYLE_BAR_CLASS, CURSOR_STYLE_UNDERLINE_CLASS, DomRendererRowFactory } from './DomRendererRowFactory';
 import { INVERTED_DEFAULT_COLOR } from '../atlas/Types';
 import { Disposable } from 'common/Lifecycle';
-import { IColorSet } from 'ui/Types';
-import { ICharSizeService } from 'ui/services/Services';
+import { IColorSet } from 'browser/Types';
+import { ICharSizeService } from 'browser/services/Services';
 
 const TERMINAL_CLASS_PREFIX = 'xterm-dom-renderer-owner-';
 const ROW_CONTAINER_CLASS = 'xterm-rows';
